ISR zwischenstand

+++ lib/memory.c
@@ -3,15 +3,15 @@
 
 //memcpy, memcmp, memmove, memset are from https://github.com/gcc-mirror/gcc/tree/master/libgcc
 
-void *memcpy (void *dest, const void *src, size_t len){
+void *memcpy (void *dest, const void *src, unsigned int len){
     char *d = dest;
     const char *s = src;
     while (len--)
         *d++ = *s++;
     return dest;
 }
-
-int memcmp (const void *str1, const void *str2, size_t count){
+/*
+int memcmp (const void *str1, const void *str2, unsigned int count){
     const unsigned char *s1 = str1;
     const unsigned char *s2 = str2;
 
@@ -23,7 +23,7 @@ int memcmp (const void *str1, const void *str2, size_t count){
     return 0;
 }
 
-void *memmove (void *dest, const void *src, size_t len){
+void *memmove (void *dest, const void *src, unsigned int len){
     char *d = dest;
     const char *s = src;
     if (d < s)
@@ -38,9 +38,10 @@ void *memmove (void *dest, const void *src, size_t len){
     return dest;
 }
 
-void *memset (void *dest, int val, size_t len){
+void *memset (void *dest, int val, unsigned int len){
     unsigned char *ptr = dest;
     while (len-- > 0)
         *ptr++ = val;
     return dest;
 }
+*/
